Move fetch actions to App component
